@@ -3,29 +3,30 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@chakra-ui/react": "^1.8.1",
-        "@emotion/react": "^11.7.1",
-        "@emotion/styled": "^11.6.0",
-        "@mui/material": "^5.10.17",
-        "@testing-library/jest-dom": "^5.16.2",
-        "@testing-library/react": "^12.1.2",
-        "@testing-library/user-event": "^13.5.0",
-        "altogic": "^2.3.8",
+        "@chakra-ui/react": "^2.4.8",
+        "@choc-ui/chakra-autocomplete": "^5.1.4",
+        "@emotion/react": "^11.10.5",
+        "@emotion/styled": "^11.10.5",
+        "@mui/material": "^5.11.4",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^14.4.3",
+        "altogic": "^2.3.9",
         "aos": "^2.3.4",
         "emailjs-com": "^3.2.0",
         "formik": "^2.2.9",
-        "material-react-table": "^1.2.10",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-icons": "^4.3.1",
-        "react-intl": "^5.24.6",
-        "react-router-dom": "^6.2.1",
-        "react-scripts": "5.0.0",
+        "material-react-table": "^1.5.8",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-icons": "^4.7.1",
+        "react-intl": "^6.2.5",
+        "react-router-dom": "^6.6.2",
+        "react-scripts": "5.0.1",
         "react-slick": "^0.29.0",
-        "react-tsparticles": "^1.42.1",
-        "sweetalert2": "^11.4.2",
-        "typewriter-effect": "^2.18.2",
-        "web-vitals": "^2.1.4",
+        "react-tsparticles": "^2.7.1",
+        "sweetalert2": "^11.7.0",
+        "typewriter-effect": "^2.19.0",
+        "web-vitals": "^3.1.1",
         "yup": "^0.32.11"
     },
     "scripts": {
