@@ -1,5 +1,4 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
@@ -8,11 +7,14 @@ import { LangProvider } from "./context/langContext";
 import { BasketProvider } from "./context/basketContext";
 import { AuthenticationProvider } from "./context/authContext";
 import { BrowserRouter as Router } from "react-router-dom";
-
+import { createRoot } from 'react-dom/client';
 import { ChakraProvider } from "@chakra-ui/react";
 import theme from "./theme";
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
   <React.StrictMode>
     <ChakraProvider theme={theme}>
       <LangProvider>
@@ -26,7 +28,6 @@ ReactDOM.render(
       </LangProvider>
     </ChakraProvider>
   </React.StrictMode>,
-  document.getElementById("root")
 );
 
 // If you want to start measuring performance in your app, pass a function
